+++ src/db/repositories/user-item-repository.ts
@@ -1,8 +1,11 @@
-import type { InsertUserItem } from '@/domain/entities/user-item'
+import type {
+  InsertUserItem,
+  SelectAllUserItems,
+  SelectUserItems,
+} from '@/domain/entities/user-item'
 import type { GetUserItemInput } from '@/domain/services/user-items/get-user-item'
-import type { GetUserItemsInput } from '@/domain/services/user-items/get-user-items'
 
-import { and, desc, eq, sql } from 'drizzle-orm'
+import { and, desc, eq, getTableColumns, lte, sql } from 'drizzle-orm'
 import { db } from '..'
 import { schema } from '../schema'
 
@@ -25,17 +28,29 @@ export async function upsertUserItem({
   )
 }
 
-export async function selectUserItems({ userId, status }: GetUserItemsInput) {
+export async function selectUserItems({
+  userId,
+  status,
+  pageSize,
+  cursor,
+}: SelectUserItems) {
   return db
     .select()
     .from(schema.userItems)
     .where(
       and(
+        cursor
+          ? lte(
+              sql`DATE_TRUNC('milliseconds', ${schema.userItems.updatedAt})`,
+              cursor
+            )
+          : undefined,
         eq(schema.userItems.userId, userId),
         eq(schema.userItems.status, status)
       )
     )
     .orderBy(desc(schema.userItems.updatedAt))
+    .limit(pageSize + 1)
 }
 
 export async function deleteUserItem(id: string) {
@@ -73,4 +88,26 @@ export async function selectUserItemStatus(userId: string) {
     .from(schema.userItems)
     .where(eq(schema.userItems.userId, userId))
     .groupBy(schema.userItems.status)
+}
+
+export async function selectAllUserItems({
+  status,
+  userId,
+}: SelectAllUserItems) {
+  const { id, tmdbId, mediaType } = getTableColumns(schema.userItems)
+
+  return db
+    .select({
+      id,
+      tmdbId,
+      mediaType,
+    })
+    .from(schema.userItems)
+    .where(
+      and(
+        eq(schema.userItems.userId, userId),
+        eq(schema.userItems.status, status)
+      )
+    )
+    .orderBy(desc(schema.userItems.updatedAt))
 }